Serious memory bugs parsing provider file
This patch represents a rework of the main parsing
providerrc function which is afftected by serious
memory leaks. The purpose of the fix is to revisit
the code base providing a few improvements and
optimizations in the way the memory structures are
allocated, used and freed.
Finally, the print function is improved and now it
allows to get some basic json like output.
The main function is reworked as well, the switch
case statement is turned into a section which can
be used only for getting input and validating it.
A new flow is then added to continue the execution
according to the cli parameters:
if -b<key> is received it has the highest priority
over the other cli options (single shot mode), or
the otp computing is run in loop mode (if '-s' is
specified).

Closes: #1
Signed-off-by: fmount <fmount@inventati.org>