Fix get_otp function and improve otp watcher
This patch fixes the way we iterate over the list of Provider and
introduces a few checks in the get_otp function. The TIME const is
increased to 30 seconds (otherwise we always compute the same thing
over and over again without a particular reason).
get_otp results also better from a memory point of view, and the
otpvalue is no longer a pointer in the NODE struct.

Signed-off-by: fmount <fmount@inventati.org>